Restrictions
The new Restrictions class represents some airspace restrictions that we
are looking to examine.

The two operations it supports are to check for violations and to deduce
new graphs to avoid the restrictions.

The next steps are to develop a search algorithm that observes these
restrictions on top of our Dijkstra search, and to read the restrictions
from an input file.